requirements.txt
pyvirtualcam~=0.9.1
mss~=6.1.0
numpy~=1.22.3
PySide6~=6.3.0
tensorflow-gpu~=2.8.0
matplotlib~=3.5.1
imageio~=2.16.1
ipython~=8.3.0